refactor(pages): extract client content for SSR metadata
- Refactor landing page to server component with metadata
- Refactor docs intro, installation, components overview pages
- Extract interactive content to separate client components
- Add proper SEO metadata and breadcrumb schemas
